% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCompartment4}
\alias{dataCompartment4}
\title{An example data set with 4 compartment}
\format{
A data frame with 21 rows and 7 variables:
\describe{
\item{temps}{vector of time}
\item{condition}{exposure concentration}
\item{replicat}{replicate of experiment}
\item{intestin}{compartiment 'intestin'}
\item{caecum}{compartiment 'caecum'}
\item{cephalon}{compartiment 'cephalon'}
\item{reste}{compartiment 'reste'}#'
}
}
\usage{
data(dataCompartment4)
}
\description{
A dataset containing an example with 4 compartments.
}
\keyword{datasets}
