% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataMaleGammarusSingle}
\alias{dataMaleGammarusSingle}
\title{An example data set with 1 compartment}
\format{
A data frame with 22 rows and 4 variables:
\describe{
\item{time}{vector of time}
\item{expw }{exposure concentration}
\item{replicate}{replicat of experiment}
\item{conc}{internal measured concentration}#'
}
}
\usage{
data(dataMaleGammarusSingle)
}
\description{
A dataset containing an example with a single compartment.
}
\keyword{datasets}
