% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFIT.R
\name{df_quant95_}
\alias{df_quant95_}
\title{Compute 95 credible intervals}
\usage{
df_quant95_(x, ...)
}
\arguments{
\item{x}{An object of class \code{fitPBK}}

\item{\dots}{Additional arguments}
}
\value{
An object of class \code{data.frame} returning median and 95 credible interval
}
\description{
Compute quantiles 95 credible intervals
}
