% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPBK.R
\name{export_interpolate}
\alias{export_interpolate}
\title{Interpolate function implemented in Stan only export for checking}
\usage{
export_interpolate(x, xpt, ypt, chain = 1, iter = 1, ...)
}
\arguments{
\item{x}{interpolation point x}

\item{xpt}{a vector of x axis (has to be same size as \code{ypt} vector)}

\item{ypt}{a vector of y axis (has to be same size as \code{ypt} vector)}

\item{chain}{number of chain}

\item{iter}{number of iteration}

\item{\dots}{Arguments passed to \code{rstan::sampling}}
}
\value{
A sample of a stanfit object returning a linear interpolation
}
\description{
This function export the linear iterpolation implemented in Stan. It
can be use to re-sample the exposure profiles.
}
