% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitPBK_C4}
\alias{fitPBK_C4}
\title{An example of fitPBK object}
\format{
A \code{fitPBK} object with:
\describe{
\item{stanPBKdata}{original data frame wrap in a formatted list for inference}
\item{stanfit}{a \code{stanfit} object resulting from inference with stan}
}
}
\usage{
data(fitPBK_C4)
}
\description{
A fitPBK object containing Bayesian inference of the
\code{dataCompartment4} data set.
}
\keyword{datasets}
