% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFIT.R
\name{plot.fitPBK}
\alias{plot.fitPBK}
\title{Plotting method for \code{fitPBK} objects}
\usage{
\method{plot}{fitPBK}(x, ...)
}
\arguments{
\item{x}{And object returned by \code{fitPBK}}

\item{\dots}{Additional arguments}
}
\value{
a plot of class \code{ggplot}
}
\description{
This is the generic \code{plot} S3 method for the
\code{fitTK}.  It plots the fit obtained for each
variable in the original dataset.
}
\examples{
# (1) load a fitPBK object
data("fitPBK_C4")
# (2) plot result of bayesian fitting
plot(fitPBK_C4)

}
