% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPPC.R
\name{ppc}
\alias{ppc}
\alias{ppc.fitPBK}
\title{Posterior predictive check plot}
\usage{
ppc(x, ...)

\method{ppc}{fitPBK}(x, ...)
}
\arguments{
\item{x}{an object used to select a method \code{ppc}}

\item{\dots}{Further arguments to be passed to generic methods}
}
\value{
a plot of class \code{ggplot}
}
\description{
Plots posterior predictive check for \code{fitPBK}
}
\examples{
# (1) load a fitPBK object
data("fitPBK_C4")
# (2) plot ppc of bayesian fitting
ppc(fitPBK_C4)


}
