% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPBK-package.R
\name{rPBK-package}
\alias{rPBK-package}
\alias{rPBK}
\title{The 'rPBK' package.}
\description{
Fit and simulate any kind of physiologically-based kinetic ('PBK') models whatever
the number of compartments
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
\keyword{"_PACKAGE"}
\keyword{internal}
